refactor (resource): move parsing logic in utility namespace
