+++ test/data/api/v2/resources/dinge.json
@@ -1,0 +1,58 @@
+{
+  "@graph" : [ {
+    "@id" : "http://rdfh.ch/0000/forbiddenResource",
+    "@type" : "knora-api:ForbiddenResource",
+    "knora-api:arkUrl" : {
+      "@type" : "xsd:anyURI",
+      "@value" : "http://ark.dasch.swiss/ark:/72163/1/0000/forbiddenResourceV"
+    },
+    "knora-api:attachedToProject" : {
+      "@id" : "http://www.knora.org/ontology/knora-admin#SystemProject"
+    },
+    "knora-api:attachedToUser" : {
+      "@id" : "http://rdfh.ch/users/root"
+    },
+    "knora-api:creationDate" : {
+      "@type" : "xsd:dateTimeStamp",
+      "@value" : "2017-10-06T11:05:37Z"
+    },
+    "knora-api:hasPermissions" : "CR knora-admin:Creator|M knora-admin:ProjectMember|V knora-admin:KnownUser|V knora-admin:UnknownUser",
+    "knora-api:userHasPermission" : "V",
+    "knora-api:versionArkUrl" : {
+      "@type" : "xsd:anyURI",
+      "@value" : "http://ark.dasch.swiss/ark:/72163/1/0000/forbiddenResourceV.20171006T110537Z"
+    },
+    "rdfs:label" : "This resource is a proxy for a resource you are not allowed to see (may depend on the context: query path)"
+  }, {
+    "@id" : "http://rdfh.ch/0001/a-thing-with-text-valuesLanguage",
+    "@type" : "anything:Thing",
+    "knora-api:arkUrl" : {
+      "@type" : "xsd:anyURI",
+      "@value" : "http://ark.dasch.swiss/ark:/72163/1/0001/a=thing=with=text=valuesLanguageB"
+    },
+    "knora-api:attachedToProject" : {
+      "@id" : "http://rdfh.ch/projects/0001"
+    },
+    "knora-api:attachedToUser" : {
+      "@id" : "http://rdfh.ch/users/9XBCrDV3SRa7kS1WwynB4Q"
+    },
+    "knora-api:creationDate" : {
+      "@type" : "xsd:dateTimeStamp",
+      "@value" : "2016-03-02T15:05:10Z"
+    },
+    "knora-api:hasPermissions" : "V knora-admin:UnknownUser|M knora-admin:ProjectMember",
+    "knora-api:userHasPermission" : "V",
+    "knora-api:versionArkUrl" : {
+      "@type" : "xsd:anyURI",
+      "@value" : "http://ark.dasch.swiss/ark:/72163/1/0001/a=thing=with=text=valuesLanguageB.20160302T150510Z"
+    },
+    "rdfs:label" : "ein verdammtes Ding mit einer Sprache"
+  } ],
+  "@context" : {
+    "rdf" : "http://www.w3.org/1999/02/22-rdf-syntax-ns#",
+    "knora-api" : "http://api.knora.org/ontology/knora-api/v2#",
+    "rdfs" : "http://www.w3.org/2000/01/rdf-schema#",
+    "xsd" : "http://www.w3.org/2001/XMLSchema#",
+    "anything" : "http://api.dasch.swiss/ontology/0001/anything/v2#"
+  }
+}
